latest from plans_console work
Did a LOT of testing and comparisons to make sure this is the latest greatest version.  See the plans_console repo for the most recent work.  Most work and updates were not done here since this directory is nested in a different repo.  Would be good to find out a better way to allow work to happen here in its correct home, while being used and tested in other repos.