fix #49
add timeout argument to add* functions that did not have it; note that a small timeout value will likely cause a timeout exception, but, the add* functions actually depend on a return value from the request so a tiny timeout value is probably not going to give the desired effect.  A full solution to do a non-blocking request while getting back a usable return value after however long it takes is probably to use futures or some other non-blocking module; and/or, to make the add* functions so that they can work without useful return values.