WIP #50
Allow mapless SartopoSession to get account data / map list.  Add openMap(mapID) to allow subsequent association with a map.  Should still go through the parts of the code that assume there is an open map, and make them die gracefully for a mapless session.  Already modified __del__ to work with a mapless session.